refactor: Move Fast*IntensityMap to org.janelia.alignment.intensity package
