+++ render-app/src/main/java/org/janelia/alignment/intensity/FastQuadraticIntensityMap.java
@@ -1,4 +1,4 @@
-package org.janelia.alignment.filter;
+package org.janelia.alignment.intensity;
 
 import net.imglib2.Dimensions;
 import net.imglib2.algorithm.blocks.AbstractBlockProcessor;
@@ -13,8 +13,8 @@ import net.imglib2.type.PrimitiveType;
 import net.imglib2.type.numeric.real.FloatType;
 import net.imglib2.util.Cast;
 import net.imglib2.util.IntervalIndexer;
-import org.janelia.alignment.filter.FastLinearIntensityMap.Coefficients;
-import org.janelia.alignment.filter.FastLinearIntensityMap.TransformCoefficients;
+import org.janelia.alignment.intensity.FastLinearIntensityMap.Coefficients;
+import org.janelia.alignment.intensity.FastLinearIntensityMap.TransformCoefficients;
 
 import java.util.Arrays;
 import java.util.function.Function;
@@ -22,7 +22,7 @@ import java.util.function.Function;
 import static net.imglib2.type.PrimitiveType.FLOAT;
 
 
-class FastQuadraticIntensityMap {
+public class FastQuadraticIntensityMap {
 
     /**
      * Apply an interpolated quadratic intensity map to blocks of the standard
